rename assets to auto_complete
