+++ app/components/primer/primer.ts
@@ -1,3 +1,3 @@
-import './auto_complete/component'
+import './auto_complete/auto_component'
 import './tab_container_component'
 import './time_ago_component'
